requirements.txt
google-api-python-client==2.1.0
selenium==3.141.0